+++ R/count.R
@@ -8,13 +8,22 @@
 #' @export
 #' @examplesIf require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)
 #' test <- polars::as_polars_df(mtcars)
+#'
+#' # grouping variables must be specified in count() and add_count()
 #' count(test, cyl)
-#'
 #' count(test, cyl, am)
-#'
 #' count(test, cyl, am, sort = TRUE, name = "count")
 #'
 #' add_count(test, cyl, am, sort = TRUE, name = "count")
+#'
+#' # tally() directly uses grouping variables of the input
+#' test |>
+#'   group_by(cyl) |>
+#'   tally()
+#'
+#' test |>
+#'   group_by(cyl, am) |>
+#'   tally(sort = TRUE, name = "count")
 count.RPolarsDataFrame <- function(x, ..., sort = FALSE, name = "n") {
   grps <- attributes(x)$pl_grps
   mo <- attributes(x)$maintain_grp_order
@@ -35,11 +44,30 @@ count.RPolarsDataFrame <- function(x, ..., sort = FALSE, name = "n") {
 
 #' @rdname count.RPolarsDataFrame
 #' @export
+tally.RPolarsDataFrame <- function(x, sort = FALSE, name = "n") {
+  grps <- attributes(x)$pl_grps
+  mo <- attributes(x)$maintain_grp_order
+  is_grouped <- !is.null(grps)
+  out <- count_(x, grps, sort = sort, name = name, new_col = FALSE)
+
+  if (length(grps) > 1) {
+    grps <- grps[-length(grps)]
+    out <- group_by(out, all_of(grps), maintain_order = mo)
+  }
+
+  add_tidypolars_class(out)
+}
+
+#' @rdname count.RPolarsDataFrame
+#' @export
 count.RPolarsLazyFrame <- count.RPolarsDataFrame
 
 #' @rdname count.RPolarsDataFrame
 #' @export
+tally.RPolarsLazyFrame <- tally.RPolarsDataFrame
 
+#' @rdname count.RPolarsDataFrame
+#' @export
 add_count.RPolarsDataFrame <- function(x, ..., sort = FALSE, name = "n") {
   grps <- attributes(x)$pl_grps
   mo <- attributes(x)$maintain_grp_order